Refactor: Combine OSGlobe and OSMap
- Fix: Fleet repair in OpsiExplore
